+++ x/starname/types/naming.go
@@ -20,9 +20,8 @@ const (
 	AttributeKeyDomainName  = "domain_name"
 	AttributeKeyAccountName = "account_name"
 	AttributeKeyDomainType  = "domain_type"
-	AttributeKeyFeePaid     = "paid_fees"
-	AttributeKeyFeePayer    = "fee_payer"
 	AttributeKeyOwner       = "owner"
+	AttributeKeyBroker      = "broker"
 
 	AttributeKeyNewCertificate          = "new_certificate"
 	AttributeKeyDeletedCertificate      = "deleted_certificate"